Split treelikes and treeadoptions into 2 routes
