--- server/routes/treeuser/treeuserValidations.js
@@ -1,9 +1,0 @@
-function validateGetTreeuser(req) {
-  if (req.query.idTree === undefined) return false;
-  if (req.query.email === undefined) return false;
-  if (req.query.request === undefined) return false;
-
-  return true;
-}
-
-module.exports = { validateGetTreeuser };
